strava_ingestion_service/pyproject.toml
[tool.poetry]
name = "strava-service"
version = "0.1.0"
description = ""
authors = ["chris <58859629+christophschaller@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "strava_service"}]

[tool.poetry.dependencies]
python = "^3.9"
sqlalchemy = "^2.0.2"
fastapi = {extras = ["all"], version = "^0.91.0"}
requests = "^2.28.2"
sentry-sdk = {extras = ["fastapi"], version = "^1.15.0"}
flet = "^0.4.0"
cryptography = "^39.0.1"


[tool.poetry.group.dev.dependencies]
database-utils = {path = "../database_utils", develop = true}
python-dotenv = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
